backend/requirements.txt
fastapi[standard]>=0.110.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.13.0
